fixed env & new svelte routing/load
