+++ sample.env
@@ -1,1 +1,1 @@
-HYGRAPH_ENDPOINT=
+VITE_HYGRAPH_ENDPOINT=
