+++ src/lib/graphql-client.js
@@ -1,4 +1,3 @@
 import { GraphQLClient } from 'graphql-request'
 const HYGRAPH_ENDPOINT = import.meta.env.VITE_HYGRAPH_ENDPOINT
-
 export const client = new GraphQLClient(HYGRAPH_ENDPOINT)
