+++ src/routes/+page.svelte
@@ -1,38 +1,11 @@
-<script context="module">
+<script>
   import Head from '$components/head.svelte'
   import ProjectCard from '$components/project-card.svelte'
-  import { client } from '$lib/graphql-client'
-  import { authorsQuery, projectsQuery } from '$lib/graphql-queries'
   import {
     authorsStore,
-    fetchAuthors,
-    fetchSiteMetadata,
     siteMetadataStore,
   } from '$stores/site-metadata'
-
-  export const load = async () => {
-    await fetchAuthors()
-    await fetchSiteMetadata()
-
-    const [authorRes, projectsRes] = await Promise.all([
-      client.request(authorsQuery),
-      client.request(projectsQuery),
-    ])
-    const { authors } = authorRes
-    const { projects } = projectsRes
-
-    return {
-      props: {
-        projects,
-        authors,
-      },
-    }
-  }
-</script>
-
-<script>
-  export let projects
-  export let authors
+  export let data
 
   const {
     siteUrl,
@@ -54,7 +27,7 @@
   Welcome to my Portfolio
 </h1>
 
-{#each authors as { name, intro, picture: { url } }}
+{#each data.authors as { name, intro, picture: { url } }}
   <div class="flex mb-40 items-end">
     <div class="mr-6">
       <h2 class="text-3xl mb-4 font-bold tracking-wider">{name}</h2>
@@ -68,7 +41,7 @@
 <div
   class="grid gap-10 md:grid-cols-4 md:px-10 lg:grid-cols-6 lg:-mx-52"
 >
-  {#each projects as { name, slug, description, image }}
+  {#each data.projects as { name, slug, description, image }}
     <ProjectCard {name} {description} url={image[0].url} {slug} />
   {/each}
 </div>
