+++ src/routes/posts/+page.svelte
@@ -1,30 +1,12 @@
-<script context="module">
+<script>
   import { page } from '$app/stores'
   import Head from '$components/head.svelte'
-  import { client } from '$lib/graphql-client'
-  import { postsQuery } from '$lib/graphql-queries'
   import {
-    fetchSiteMetadata,
-    siteMetadataStore,
+      siteMetadataStore
   } from '$stores/site-metadata'
   import { marked } from 'marked'
   import { onMount } from 'svelte'
-
-  export const load = async () => {
-    await fetchSiteMetadata()
-
-    const { posts } = await client.request(postsQuery)
-
-    return {
-      props: {
-        posts,
-      },
-    }
-  }
-</script>
-
-<script>
-  export let posts
+  export let data
   let pathname
 
   onMount(async () => {
@@ -47,7 +29,7 @@
 
 <h1 class="text-4xl mb-10 font-extrabold">Blog posts</h1>
 
-{#each posts as { title, slug, content, coverImage, tags }}
+{#each data.posts as { title, slug, content, coverImage, tags }}
   <div class="card text-center shadow-2xl mb-20">
     <figure class="">
       <img
