+++ src/routes/projects/+page.svelte
@@ -1,30 +1,13 @@
-<script context="module">
+<script>
   import { page } from '$app/stores'
   import Head from '$components/head.svelte'
   import ProjectCard from '$components/project-card.svelte'
-  import { client } from '$lib/graphql-client'
-  import { projectsQuery } from '$lib/graphql-queries'
   import {
-    fetchSiteMetadata,
-    siteMetadataStore,
+      siteMetadataStore
   } from '$stores/site-metadata'
   import { onMount } from 'svelte'
 
-  export const load = async () => {
-    await fetchSiteMetadata()
-
-    const { projects } = await client.request(projectsQuery)
-
-    return {
-      props: {
-        projects,
-      },
-    }
-  }
-</script>
-
-<script>
-  export let projects
+  export let data
   let pathname
 
   onMount(async () => {
@@ -52,7 +35,7 @@
 <div
   class="grid gap-10 md:grid-cols-4 md:px-10 lg:grid-cols-6 lg:-mx-52"
 >
-  {#each projects as { name, slug, description, image }, index}
+  {#each data.projects as { name, slug, description, image }, index}
     <ProjectCard
       {name}
       {description}
