+++ src/routes/projects/[slug]/+page.svelte
@@ -1,32 +1,12 @@
-<script context="module">
+<script>
   import { page } from '$app/stores'
   import Head from '$components/head.svelte'
-  import { client } from '$lib/graphql-client'
-  import { projectQuery } from '$lib/graphql-queries'
   import {
-    fetchSiteMetadata,
-    siteMetadataStore,
+      siteMetadataStore
   } from '$stores/site-metadata'
   import { marked } from 'marked'
   import { onMount } from 'svelte'
-
-  export const load = async ({ params }) => {
-    await fetchSiteMetadata()
-
-    const { slug } = params
-    const variables = { slug }
-    const { project } = await client.request(projectQuery, variables)
-
-    return {
-      props: {
-        project,
-      },
-    }
-  }
-</script>
-
-<script>
-  export let project
+  export let data
   let pathname
 
   onMount(async () => {
@@ -41,8 +21,8 @@
 </script>
 
 <Head
-  title={`${project.name} · ${siteName}`}
-  description={project.description.slice(0, 120)}
+  title={`${data.project.name} · ${siteName}`}
+  description={data.project.description.slice(0, 120)}
   image={openGraphDefaultImage.url}
   url={`${siteUrl}${pathname}`}
 />
@@ -50,17 +30,17 @@
 <div class="sm:-mx-5 md:-mx-10 lg:-mx-20 xl:-mx-38 mb-5">
   <img
     class="rounded-lg"
-    src={project.image[0].url}
-    alt={project.title}
+    src={data.project.image[0].url}
+    alt={data.project.title}
   />
 </div>
 
-<h1 class="text-4xl font-semibold mb-5">{project.name}</h1>
+<h1 class="text-4xl font-semibold mb-5">{data.project.name}</h1>
 
 <div class="mb-5 flex justify-between">
   <div>
-    {#if project.tags}
-      {#each project.tags as tag}
+    {#if data.project.tags}
+      {#each data.project.tags as tag}
         <span
           class="badge badge-primary mr-2 hover:bg-primary-focus cursor-pointer"
           >{tag}</span
@@ -73,10 +53,10 @@
 <div
   class="mb-5 prose flex prose-a:text-primary hover:prose-a:text-primary-focus"
 >
-  <a class="mr-5" href={project.demo}>Demo</a>
-  <a href={project.sourceCode}>Source Code</a>
+  <a class="mr-5" href={data.project.demo}>Demo</a>
+  <a href={data.project.sourceCode}>Source Code</a>
 </div>
 
 <article class="prose prose-xl">
-  {@html marked(project.description)}
+  {@html marked(data.project.description)}
 </article>
