+++ src/stores/site-metadata.js
@@ -12,18 +12,15 @@ export const siteMetadataStore = writable([])
 
 export const fetchSocials = async () => {
   const { socials } = await client.request(socialsQuery)
-
   return socialsStore.set(socials[0])
 }
 
 export const fetchAuthors = async () => {
   const { authors } = await client.request(authorsQuery)
-
   return authorsStore.set(authors[0])
 }
 
 export const fetchSiteMetadata = async () => {
   const { projectMetadatas } = await client.request(siteMetadataQuery)
-
   return siteMetadataStore.set(projectMetadatas[0])
 }
